+++ src/lib/pages/tx-details/components/evm-related-tx-section/index.tsx
@@ -1,14 +1,13 @@
-import { Flex, Heading } from "@chakra-ui/react";
+import { Flex, Heading, Text } from "@chakra-ui/react";
 
 import { useMobile } from "lib/app-provider";
 import { EvmMethodChip } from "lib/components/EvmMethodChip";
-import { EvmToCell } from "lib/components/EvmToCell";
 import { ExplorerLink } from "lib/components/ExplorerLink";
 import { CustomIcon } from "lib/components/icon";
 import { Loading } from "lib/components/Loading";
 import { EmptyState } from "lib/components/state";
 import { useTxDataJsonRpc } from "lib/services/tx";
-import { formatEvmTxHash } from "lib/utils";
+import { formatEvmTxHash, getEvmToAddress } from "lib/utils";
 
 import { EvmRelatedField } from "./EvmRelatedField";
 
@@ -29,6 +28,8 @@ const EvmRelatedTxSectionBody = ({ evmTxHash }: EvmRelatedTxSectionProps) => {
         py={0}
       />
     );
+
+  const toAddress = getEvmToAddress(data);
 
   return (
     <>
@@ -53,14 +54,20 @@ const EvmRelatedTxSectionBody = ({ evmTxHash }: EvmRelatedTxSectionProps) => {
         name="arrow-right"
         boxSize={5}
         color="gray.600"
-        mt={2}
         display={{ base: "none", xl: "block" }}
       />
       <EvmRelatedField label="To">
-        <EvmToCell
-          to={data.tx.to}
-          contractAddress={data.txReceipt.contractAddress}
-        />
+        {toAddress ? (
+          <ExplorerLink
+            type={toAddress.type}
+            value={toAddress.address}
+            showCopyOnHover
+          />
+        ) : (
+          <Text variant="body2" color="text.disabled">
+            -
+          </Text>
+        )}
       </EvmRelatedField>
     </>
   );
@@ -69,13 +76,13 @@ const EvmRelatedTxSectionBody = ({ evmTxHash }: EvmRelatedTxSectionProps) => {
 export const EvmRelatedTxSection = ({
   evmTxHash,
 }: EvmRelatedTxSectionProps) => (
-  <Flex direction="column" gap={4} mb={8} minW="fit-content">
+  <Flex direction="column" gap={4} mb={8} minW="330px">
     <Heading as="h6" variant="h6">
       Related EVM Transaction
     </Heading>
     <Flex
       direction={{ base: "column", xl: "row" }}
-      alignItems={{ base: "start", xl: "center" }}
+      alignItems={{ base: "start", xl: "end" }}
       gap={4}
       border="1px solid var(--chakra-colors-gray-700)"
       borderRadius="8px"